+++ blocklets/ai-studio/src/components/project-settings/avatar.tsx
@@ -14,7 +14,6 @@ export default function ProjectSettingsAvatar({ value, onChange }: { value: stri
 
   const handleOpen = () => {
     const allowedFileTypes = ['image/png', 'image/jpeg', 'image/jpg'];
-    // @ts-ignore
     const uploader = uploaderRef?.current?.getUploader();
     if (uploader?.opts?.restrictions?.allowedFileTypes) {
       uploader.opts.restrictions.allowedFileTypes = allowedFileTypes;