+++ packages/code-editor/src/libs/type.tsx
@@ -1,0 +1,1 @@
+export type EditorInstance = ReturnType<(typeof import('monaco-editor'))['editor']['create']>;