+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.37 (2024-8-26)
+
+- feat: support editor emmet plugin & theme
+
+## 0.4.36 (2024-8-26)
+
+- fix: skip xss middleware for some internal api
+
+## 0.4.35 (2024-8-22)
+
+- chore: use @blocklet/uploader to uploader file
+
 ## 0.4.34 (2024-8-22)
 
 - fix: skip non-string parameter rendering